Revert "sdfsfdsdf"
This reverts commit d4f9d6352a5c4dffcfbc9e5f7c44517b412dcf38.

+++ app/src/main/java/com/snow/dreamdiary/feature_dailysurvey/domain/model/DailySurveyData.kt
@@ -8,7 +8,7 @@ import androidx.room.PrimaryKey
 data class DailySurveyData(
     val didDream: Boolean = false,
     val dreamsNum: Int = -1,
-    val timeSlept: Int = 8,
+    val timeSlept: Long = 8,
     @IntRange(from = 0, to = 10) val health: Int = 5,
     @IntRange(from = 0, to = 10) val physicalActivity: Int = 5,
     @IntRange(from = 0, to = 10) val comfortness: Int = 5,
